[Interpreter] Added threaded dispatch

+++ interpreter.h
@@ -8,18 +8,19 @@
 class Interpreter final {
 public:
     Interpreter();
-
     void loadProgram(const std::string &filename);
     void executeProgram();
     ~Interpreter();
 
 private:
-    std::vector<int> registers;
-    std::vector<int> program;
+    std::vector<uint64_t> registers;
+    std::vector<uint64_t> program;
     int accumulator;
     const int regNum = 8;
 
-    void executeInstruction(int &pc);
+    void executeInstruction(uint64_t &pc);
+
+    // void executeInstruction(int &pc);
 };
 
 #endif // INTERPRETER_H
